requirements.txt
# numpy==1.26.4
numpy
# pandas==1.4.0
pandas
packaging
wheel
ninja
sentence_transformers==3.1.1
fastapi[standard]
uvicorn[standard]
pyodbc
python-dotenv
healthcheck-python
datasets
accelerate
docx
# underthesea==6.8.4
# vinorm==2.0.7
# pydub==0.25.1
# deepspeed
# SQLAlchemy
transformers==4.45.1
peft
bitsandbytes
trl
streamlit
ipywidgets
# dill==0.3.9
# fsspec==2024.6.1
# -e git+https://github.com/thinhlpg/TTS.git@ff217b3f27b294de194cc59c5119d1e08b06413c#egg=TTS


# pip install -r requirements.txt --use-deprecated=legacy-resolver
# pip install flash-attn --no-build-isolation
# numpy
# pandas
# huggingface-hub
# huggingface-hub==0.25.1
# pydantic
# pydantic==2.9.2
# pydantic_core==2.23.4

# openai
# unidecode
# Unidecode==1.3.8

# soundfile==0.12.1

# librosa==0.10.2.post1

# underthesea
# vinorm
# soundfile
# pydub
# librosa

# numpy
# pandas
# sentence_transformers
# fastapi[standard]
# uvicorn[standard]
# pyodbc
# python-dotenv
# huggingface-hub
# pydantic
# healthcheck-python
# openai
# unidecode
# datasets
# accelerate
# docx
# datasets



#torch
#torchvision
#torchaudio
#spacy
#transformers
#accelerate
#bitsandbytes
#jupyter
#wheel
#huggingface_hub
#packaging
#inja
#autoawq
#flash-attn --no-build-isolation
#Requests
#PyMuPDF==1.23.26
#matplotlib